add masking to ESMC_FieldRegridStore() and lots of examples of source masking in serial mode

+++ src/Infrastructure/Grid/tests/ESMC_GridUTest.C
@@ -1,4 +1,4 @@
-// $Id: ESMC_GridUTest.C,v 1.4 2012/04/03 03:11:27 theurich Exp $
+// $Id: ESMC_GridUTest.C,v 1.5 2012/04/04 16:58:24 rokuingh Exp $
 //
 // Earth System Modeling Framework
 // Copyright 2002-2012, University Corporation for Atmospheric Research,
@@ -258,9 +258,9 @@ int main(void){
   p = 0;
   for (int i1=exLBound[1]; i1<=exUBound[1]; ++i1) {
     for (int i0=exLBound[0]; i0<=exUBound[0]; ++i0) {
-      gridMask[p]=(int)(i0);
+      gridMask[p]=i0;
       printf("PET%d - set gridMask[%d] = %d (%d)\n", localPet, p, 
-        (int)(i0), gridMask[p]);
+        i0, gridMask[p]);
       ++p;
     }
   }
